Extracted image pre-loading into its own module
