+++ src/additional.js
@@ -1,5 +1,6 @@
 import decks from './decks'
 import events from './events'
+import imageLoader from './imageLoader'
 import navigation from './navigation'
 import spinner from './spinner'
 import storage from './storage'
@@ -10,70 +11,45 @@ require('es6-promise').polyfill();
 require('../stylesheet/additional.scss');
 
 
+// disable native touchmove behavior to prevent overscroll
+$(document).on('touchmove', event => event.preventDefault());
+
+// handles closing browser's tab/window or navigating away from MJ
+$(window).on('unload beforeunload', function() {
+    events.trigger('exitApp', null, true);
+});
+
 /**
  * List of images to load before leaving the splash screen.
  * @type {string[]}
  */
-var imagesToLoad = [
+let images = [
     'images/jewel.svg'
 ];
 
-(function setup() {
-    // disable native touchmove behavior to prevent overscroll
-    $(document).on('touchmove', event => event.preventDefault());
-
-    // handles closing browser's tab/window or navigating away from MJ
-    $(window).on('unload beforeunload', function() {
-        events.trigger('exitApp', null, true);
-    });
-
-    let promises = [];
-
-    // initializes the storage
-    promises.push(storage.setup());
-
-    // loads all images
-    promises.push(loadImages().catch(function (error) {
-        // Well, that's life. Hopefully everything still works without images.
-        console.error(error);
-    }));
-
-    // once everything is loaded and initialized, let the user proceed
-    Promise.all(promises).then(function() {
-        var $screen = $('#splash-screen');
-        $screen.click(function () {
-            $screen.hide();
-            if (decks.getSelectedDeck() == null) {
-                navigation.navigateTo('settings');
-            } else {
-                navigation.navigateTo('main-menu');
-            }
-        });
-        $screen.find('.continue').removeClass('hidden');
-        spinner.stop();
-    });
-})();
-
 /**
- * Returns a promise which fulfills when all images are loaded.
- * @returns {Promise}
+ * List of promises which need to be fulfilled before hiding the splash screen.
+ * @type {Promise[]}
  */
-function loadImages() {
-    let promises = imagesToLoad.map(url => loadImage(url).catch(() => console.error(`failed to load ${url}`)));
-    return Promise.all(promises).then(() => console.log('images loaded'));
-}
+let promises = [];
 
-/**
- * Returns a promise which fulfills when the specified image finishes loading.
- * @param {string} url
- * @returns {Promise}
- */
-function loadImage(url) {
-    console.log(`loading "${url}"...`);
-    return new Promise(function (resolve, reject) {
-        let image = new Image();
-        image.onload = resolve;
-        image.onerror = reject;
-        image.src = url;
+// initializes the storage
+promises.push(storage.setup());
+
+// loads all images
+promises.push(imageLoader.loadImages(images).then(() => console.log('images loaded')));
+
+// once everything is loaded and initialized, let the user proceed
+Promise.all(promises).then(function() {
+    var $screen = $('#splash-screen');
+    $screen.click(function() {
+        $screen.hide();
+        if (decks.getSelectedDeck() == null) {
+            navigation.navigateTo('settings');
+        } else {
+            navigation.navigateTo('main-menu');
+        }
     });
-}
+    $screen.find('.continue').removeClass('hidden');
+    spinner.stop();
+});
